requirements.txt
asgiref==3.7.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2023.10
dj-database-url==2.1.0
Django==5.0.1
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-froala-editor==4.1.4
django-js-asset==2.2.0
django-quill-editor==0.1.40
gunicorn==21.2.0
h11==0.14.0
mysql-connector-python==8.2.0
mysqlclient==2.2.1
packaging==24.0
pillow==10.2.0
protobuf==4.21.12
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2023.4
uvicorn==0.29.0
waitress==3.0.0
whitenoise==6.6.0
